BUG/MINOR: cli: properly decrement ref count on tables during failed dumps
Gernot Pörner reported some constant leak of ref counts for stick tables
entries. It happens that this leak was not at all in the regular traffic
path but on the "show table" path. An extra ref count was taken during
the dump if the output had to be paused, and it was released upon clean
termination or an error detected in the I/O handler. But the release
handler didn't do it, while it used to properly do it for the sessions
dump.

This fix needs to be backported to 1.6.

(cherry picked from commit 2fe1b921639fd152092cead0fe0398ef0b15c5a3)
(cherry picked from commit 7ad199f8dda7b2688ecbd6ceb57e210132822824)